requirements.txt
asgiref==3.5.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
defusedxml==0.7.1
Django==4.0.6
django-cors-headers==3.13.0
django-extensions==3.2.0
djangorestframework==3.13.1
idna==3.3
jwt==1.3.1
mysqlclient==2.1.1
oauthlib==3.2.0
py==1.11.0
pycparser==2.21
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2022.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.11
